It is working again. This is a good state. It still has a strane tick of making lower left point very negative.

+++ _utilities/kernels/k_equal.m
@@ -1,3 +1,3 @@
 function k = k_equal(b)    
-    k = @(x1,x2) b(x1,x2) == 0;
+    k = @(x1,x2) double(b(x1,x2) == 0);
 end
